Multistage Boss pt. 1

+++ Team_DDM/Assets/Scripts/enemyStats.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 enum enemies
 {
-    basicRange = 0, basicMelee, bulletHell, spider
+    basicRange = 0, basicMelee, bulletHell, boss, spider
 }// remove spider code eventually
 
 public class enemyStats : MonoBehaviour, IDamage
@@ -16,12 +16,15 @@ public class enemyStats : MonoBehaviour, IDamage
     [SerializeField] GameObject gunToDrop;
     [SerializeField] GameObject ammoToDrop;
     [SerializeField] GameObject healthToDrop;
+    [SerializeField] float enemyInvTimer;
 
     enemyAI aiScript;
     Animator anim;
     int enemyTypeID;
     CapsuleCollider collider;
-    
+    bool isInv = false;
+    bool phase1, phase2, phase3 = false;
+
 
     // Start is called before the first frame update
     void Start()
@@ -32,7 +35,7 @@ public class enemyStats : MonoBehaviour, IDamage
         aiScript = GetComponent<enemyAI>();
         enemyTypeID = aiScript.GetEnemyTypeID();
         collider = GetComponent<CapsuleCollider>();
-        
+
     }
 
 
@@ -43,29 +46,38 @@ public class enemyStats : MonoBehaviour, IDamage
     /// <param name="dmg"></param>
     public void takeDamage(int dmg)
     {
-        HP -= dmg;
-        Debug.Log(this.gameObject.name + "took damage");
-        //aiScript.GetHeadPos().gameObject.GetComponent<Collider>().enabled = false;
-        if (HP <= 0)
-        {
-            gameManager.instance.RoomFinished(-1);
-            DropItems();
-            if ((enemies)enemyTypeID == enemies.spider)
+        if (!isInv)// not invincible
+        { 
+            HP -= dmg;
+            Debug.Log(this.gameObject.name + "took damage");
+            //aiScript.GetHeadPos().gameObject.GetComponent<Collider>().enabled = false;
+            if (HP <= 0)
             {
-                Destroy(gameObject);// kill enemy
+                gameManager.instance.RoomFinished(-1);
+                DropItems();
+                if ((enemies)enemyTypeID == enemies.spider)
+                {
+                    Destroy(gameObject);// kill enemy
+                }
+                else
+                {
+                    anim.SetBool("Dead", true);
+                    aiScript.TurnOffNavMesh();
+                    collider.enabled = false;
+                }
             }
             else
             {
-                anim.SetBool("Dead", true);
-                aiScript.TurnOffNavMesh();
-                collider.enabled = false;
+                anim.SetTrigger("Hit");
+                StartCoroutine(flashEnemyDamage());
+            }
+
+            if (enemyTypeID == (int)enemies.boss)// run enemy boss functions
+            {
+                EnemyBossPhaseEnter();
             }
         }
-        else
-        {
-            anim.SetTrigger("Hit");
-            StartCoroutine(flashEnemyDamage());
-        }
+
     }
 
     /// <summary>
@@ -96,5 +108,36 @@ public class enemyStats : MonoBehaviour, IDamage
         }
     }
 
+    /// <summary>
+    /// checks the enemy health and has it enter the selected enemy boss phase
+    /// </summary>
+    void EnemyBossPhaseEnter()
+    {
+        if (HP > (HPOrig * .7f) && !phase1)// phase 1(melee)
+        {
+            phase1 = true;
+            StartCoroutine(InvincibilityFrames());
+            anim.SetTrigger("Phase 1");
+        }
+        else if (HP > (HPOrig * .5f) && !phase2)// phase 2(single shot)
+        {
+            phase2 = true;
+            StartCoroutine(InvincibilityFrames());
+            anim.SetTrigger("Phase 2");
+        }
+        else if (!phase3)// phase 3(bullet hell boss)
+        {
+            phase3 = true;
+            StartCoroutine(InvincibilityFrames());
+            anim.SetTrigger("Phase 3");
+        }
+    }
+
+    IEnumerator InvincibilityFrames()
+    {
+        isInv = true;
+        yield return new WaitForSeconds(enemyInvTimer);
+        isInv = false;
+    }
 
 }
